requirements.txt
Flask==2.2.5
beautifulsoup4 
bs4
frozendict
geopy
ipython
kafka-python
matplotlib
numpy
py4j
pyelasticsearch
pymongo
requests
scipy
selenium
scikit-learn
tabulate
tldextract
wikipedia
findspark
iso8601
urllib3==1.24
joblib




Werkzeug==2.2.3
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.8.2
eventlet==0.33.3

cassandra-driver



